Add i18n-enabled Jinja2 plugin and add custom commands option
The i18n-enabled Jinja2 plugin is working and usable, but now that a working
prototype has been established we should try refactoring it into a plugin
that does only i18n. The problem with that, unfortunately, is that we do not
want this i18n plugin to be applied to JSON for example, and we also need to
insert variables into the Jinja2 templates for this to work. Jinja2 and i18n
is therefore tightly coupled, and hard to separate right now.